add java. improve swift
