+++ 3-sum-of-digits-to-power/Swift/sum-of-digits-to-power/sum-of-digits-to-power/Package.swift
@@ -1,0 +1,22 @@
+// swift-tools-version: 5.7
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "sum-of-digits-to-power",
+    dependencies: [
+        // Dependencies declare other packages that this package depends on.
+        // .package(url: /* package url */, from: "1.0.0"),
+    ],
+    targets: [
+        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
+        // Targets can depend on other targets in this package, and on products in packages this package depends on.
+        .executableTarget(
+            name: "sum-of-digits-to-power",
+            dependencies: []),
+        .testTarget(
+            name: "sum-of-digits-to-powerTests",
+            dependencies: ["sum-of-digits-to-power"]),
+    ]
+)
