+++ 3-sum-of-digits-to-power/Swift/sum-of-digits-to-power/sum-of-digits-to-power/Tests/sum-of-digits-to-powerTests/sum_of_digits_to_powerTests.swift
@@ -1,0 +1,11 @@
+import XCTest
+@testable import sum_of_digits_to_power
+
+final class sum_of_digits_to_powerTests: XCTestCase {
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
+        XCTAssertEqual(sum_of_digits_to_power().text, "Hello, World!")
+    }
+}
