+++ 3-sum-of-digits-to-power/java/Sumdigits.java
@@ -1,0 +1,10 @@
+// class Declaration
+public class Sumdigits
+{
+
+
+   public static void main(String[] args)
+   {
+       System.out.println("Narcissistic Number");
+   }
+}
