+++ 3-sum-of-digits-to-power/python/digitsum.py
@@ -19,7 +19,7 @@ def max_num(exponent):
 def sum_digits_rec(num):
     if num in cache:
         return cache[num]
-    cache[num] = sum_digits_rec(num // 10) + cache[num % 10]
+    cache[num] = sum_digits_rec(num // 10) + sum_digits_rec(num % 10)
     return cache[num]
